#53 10-learning-videos/
optimize the directory name of the "learning videos".

The "learning videos" are documented because of their information density and high quality
So, these videos need to be put in their own directories for separation and clarity.
So, the learning notes and learning examples can sit close to each other for future reference.